reorganise post merge
