--- src/super_gradients/common/auto_logging/__init__.py
@@ -1,4 +1,0 @@
-from super_gradients.common.auto_logging.auto_logger import AutoLoggerConfig
-from super_gradients.common.auto_logging.console_logging import ConsoleSink
-
-__all__ = ["AutoLoggerConfig", "ConsoleSink"]
