+++ src/super_gradients/common/environment/ddp_utils.py
@@ -2,13 +2,11 @@ import os
 import socket
 from functools import wraps
 
-from super_gradients.common.environment.environment_config import device_config
-from super_gradients.common.environment.argparse_utils import pop_arg
-from super_gradients.common.environment.omegaconf_utils import register_hydra_resolvers
+from super_gradients.common.environment.device_utils import device_config
+from super_gradients.common.abstractions.abstract_logger import get_logger
 
 
-DDP_LOCAL_RANK = int(os.getenv("LOCAL_RANK", default=-1))
-INIT_TRAINER = False
+logger = get_logger(__name__)
 
 
 def init_trainer():
@@ -16,28 +14,14 @@ def init_trainer():
     Initialize the super_gradients environment.
 
     This function should be the first thing to be called by any code running super_gradients.
-    It resolves conflicts between the different tools, packages and environments used and prepares the super_gradients environment.
     """
-    global INIT_TRAINER, DDP_LOCAL_RANK
-
-    if not INIT_TRAINER:
-        register_hydra_resolvers()
-
-        # We pop local_rank if it was specified in the args, because it would break
-        args_local_rank = pop_arg("local_rank", default_value=-1)
-
-        # Set local_rank with priority order (env variable > args.local_rank > args.default_value)
-        DDP_LOCAL_RANK = int(os.getenv("LOCAL_RANK", default=args_local_rank))
-        INIT_TRAINER = True
+    logger.info("You are using init_trainer which is not required anymore.")
+    return
 
 
 def is_distributed() -> bool:
-    return DDP_LOCAL_RANK >= 0
-
-
-def is_rank_0() -> bool:
-    """Check if the node was launched with torch.distributed.launch and if the node is of rank 0"""
-    return os.getenv("LOCAL_RANK") == "0"
+    """Check if current process is a DDP subprocess."""
+    return device_config.assigned_rank >= 0
 
 
 def is_launched_using_sg():
