+++ src/super_gradients/common/environment/omegaconf_utils.py
@@ -36,3 +36,6 @@ def register_hydra_resolvers():
     OmegaConf.register_new_resolver("getitem", lambda container, key: container[key], replace=True)  # get item from a container (list, dict...)
     OmegaConf.register_new_resolver("first", lambda lst: lst[0], replace=True)  # get the first item from a list
     OmegaConf.register_new_resolver("last", lambda lst: lst[-1], replace=True)  # get the last item from a list
+
+
+register_hydra_resolvers()
