+++ tests/unit_tests/train_logging_test.py
@@ -1,6 +1,6 @@
 import unittest
 from super_gradients import Trainer
-from super_gradients.common.auto_logging import AutoLoggerConfig
+from super_gradients.common.auto_logging.auto_logger import AutoLoggerConfig
 from super_gradients.training.dataloaders.dataloaders import classification_test_dataloader
 from super_gradients.training.metrics import Accuracy, Top5
 from super_gradients.training.models import ResNet18
